rename file board -> boards
